Refactored schedule screen, instead of using same props, added on top of the existing ones in types to reuse code. after changing that, needed to fix the way names are being loaded in schedule/results
