+++ app/Styles/StandingsStyles.tsx
@@ -1,0 +1,26 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+    safeArea: {
+      flex: 1,
+      backgroundColor: '#ffffff',
+    },
+    switchContainer: {
+      flexDirection: 'row',
+      alignItems: 'center',
+      paddingHorizontal: 16,
+      paddingVertical: 8,
+    },
+    switchLabel: {
+      fontSize: 16,
+      marginRight: 8,
+    },
+    loadingContainer: {
+      marginTop: 50,
+      alignItems: 'center',
+    },
+    loadingText: {
+      marginTop: 10,
+      fontSize: 16,
+    },
+  });
